+++ model/src/guild/guild.rs
@@ -53,7 +53,7 @@ pub struct Guild {
         serialize_with = "Snowflake::serialize_option_to_int"
     )]
     pub system_channel_id: Option<Snowflake>,
-    pub system_channels_flags: Option<u8>,
+    pub system_channels_flags: Option<u32>,
     pub rules_channel_id: Option<Snowflake>,
     pub joined_at: Option<DateTime<Utc>>,
     pub large: Option<bool>,
